+++ tests/VObject/ITip/BrokerAttendeeReplyTest.php
@@ -268,14 +268,16 @@ ICS;
 
     }
 
+    /**
+     * In this test, an event exists in an attendees calendar. The event
+     * is recurring, and the attendee deletes 1 instance of the event.
+     * This instance shows up in EXDATE
+     *
+     * This should automatically generate a DECLINED message for that
+     * specific instance.
+     */
     function testCreateReplyByException() {
 
-        // In this test, an event exists in an attendees calendar. The event
-        // is recurring, and the attendee deletes 1 instance of the event.
-        // This instance shows up in EXDATE
-        //
-        // This should automatically generate a DECLINED message for that
-        // specific instance.
 
         $oldMessage = <<<ICS
 BEGIN:VCALENDAR
@@ -326,6 +328,76 @@ BEGIN:VEVENT
 UID:foobar
 SEQUENCE:1
 RECURRENCE-ID:20140818T200000Z
+ORGANIZER:mailto:organizer@example.org
+ATTENDEE;PARTSTAT=DECLINED:mailto:one@example.org
+END:VEVENT
+END:VCALENDAR
+ICS
+
+            ),
+        );
+        $result = $this->parse($oldMessage, $newMessage, $expected);
+
+    }
+
+    /**
+     * This test is identical to the last, but now we're working with
+     * timezones.
+     *
+     * @depends testCreateReplyByException
+     */
+    function testCreateReplyByExceptionTz() {
+
+
+        $oldMessage = <<<ICS
+BEGIN:VCALENDAR
+VERSION:2.0
+BEGIN:VEVENT
+UID:foobar
+SEQUENCE:1
+DTSTART;TZID=America/Toronto:20140811T200000
+RRULE:FREQ=WEEKLY
+ORGANIZER:mailto:organizer@example.org
+ATTENDEE:mailto:one@example.org
+END:VEVENT
+END:VCALENDAR
+ICS;
+
+        $newMessage = <<<ICS
+BEGIN:VCALENDAR
+VERSION:2.0
+BEGIN:VEVENT
+UID:foobar
+SEQUENCE:1
+DTSTART;TZID=America/Toronto:20140811T200000
+RRULE:FREQ=WEEKLY
+ORGANIZER:mailto:organizer@example.org
+ATTENDEE:mailto:one@example.org
+EXDATE;TZID=America/Toronto:20140818T200000
+END:VEVENT
+END:VCALENDAR
+ICS;
+
+        $version = \Sabre\VObject\Version::VERSION;
+        $expected = array(
+            array(
+                'uid' => 'foobar',
+                'method' => 'REPLY',
+                'component' => 'VEVENT',
+                'sender' => 'mailto:one@example.org',
+                'senderName' => null,
+                'recipient' => 'mailto:organizer@example.org',
+                'recipientName' => null,
+                'message' => <<<ICS
+BEGIN:VCALENDAR
+VERSION:2.0
+PRODID:-//Sabre//Sabre VObject $version//EN
+CALSCALE:GREGORIAN
+METHOD:REPLY
+BEGIN:VEVENT
+UID:foobar
+SEQUENCE:1
+RECURRENCE-ID;TZID=America/Toronto:20140818T200000
 ORGANIZER:mailto:organizer@example.org
 ATTENDEE;PARTSTAT=DECLINED:mailto:one@example.org
 END:VEVENT