+++ ChangeLog.md
@@ -8,6 +8,7 @@ ChangeLog
   setters on properties. (`$event->DTSTART = new DateTime('now')`).
 * #111: iTip Broker does not process attendee adding events to EXDATE.
 * #112: EventIterator now sets TZID on RECURRENCE-ID.
+* #113: Timezone support during creation of iTip REPLY messages.
 
 
 3.3.0 (2014-08-07)